eliminated obsolete addClient form
